blog/unpacking_elfuck_packer/assets/scripts/requirements.txt
qiling==1.4.5
pyelftools==0.29